# api-specs/openapi.yaml
openapi: 3.1.0
info:
  title: GPU Inference Platform API
  version: 0.1.0
  description: |
    Client and operational HTTP contracts.
    Implementation status: not served (Session 3 foundation only).
    Prose and validation rules: docs/contracts/openai-api.md
    Typed models: packages/common-schemas (gpu-inference-common-schemas).
  x-implementation-status: not-started
  x-session: "3"
servers:
  - url: http://localhost:8080
    description: Local gateway (planned)
security:
  - bearerAuth: []
paths:
  /v1/chat/completions:
    post:
      operationId: createChatCompletion
      summary: Create chat completion
      tags: [inference]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Completion or SSE stream
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
            text/event-stream:
              schema:
                type: string
                description: SSE chunks; see ChatCompletionChunk
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '502':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '504':
          $ref: '#/components/responses/Error'
  /v1/completions:
    post:
      operationId: createCompletion
      summary: Create legacy text completion
      tags: [inference]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompletionRequest'
      responses:
        '200':
          description: Completion or SSE stream
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextCompletionResponse'
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /health:
    get:
      operationId: health
      summary: Liveness
      security: []
      tags: [operations]
      responses:
        '200':
          description: Process alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /ready:
    get:
      operationId: ready
      summary: Readiness
      security: []
      tags: [operations]
      responses:
        '200':
          description: Ready to serve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
        '503':
          description: Not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
  /version:
    get:
      operationId: version
      summary: Service version
      security: []
      tags: [operations]
      responses:
        '200':
          description: Build identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionResponse'
  /metrics:
    get:
      operationId: metrics
      summary: Prometheus metrics
      security: []
      tags: [operations]
      responses:
        '200':
          description: Prometheus text exposition
          content:
            text/plain:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ReadyResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ready, not_ready]
        reason:
          type: string
    VersionResponse:
      type: object
      required: [service, version, gateway_instance_id]
      properties:
        service:
          type: string
        version:
          type: string
        gateway_instance_id:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, message, code]
          properties:
            type:
              type: string
            message:
              type: string
            code:
              type: string
            param:
              type: string
              nullable: true
            retry_after_ms:
              type: integer
              minimum: 0
    ChatCompletionRequest:
      type: object
      additionalProperties: false
      required: [model, messages]
      properties:
        model:
          type: string
        messages:
          type: array
          minItems: 1
          maxItems: 128
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
                enum: [system, user, assistant]
              content:
                type: string
        stream:
          type: boolean
          default: false
        max_tokens:
          type: integer
          minimum: 1
        temperature:
          type: number
          minimum: 0
          maximum: 2
        top_p:
          type: number
          minimum: 0
          maximum: 1
    ChatCompletionResponse:
      type: object
      required: [id, object, created, model, choices]
      properties:
        id:
          type: string
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: string
                enum: [stop, length, cancelled, error]
        usage:
          $ref: '#/components/schemas/Usage'
    ChatCompletionChunk:
      type: object
      description: One SSE data line JSON payload
      properties:
        id:
          type: string
        object:
          type: string
          enum: [chat.completion.chunk]
        choices:
          type: array
          items:
            type: object
            properties:
              delta:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: string
                nullable: true
    CompletionRequest:
      type: object
      additionalProperties: false
      required: [model, prompt]
      properties:
        model:
          type: string
        prompt:
          type: string
        stream:
          type: boolean
          default: false
        max_tokens:
          type: integer
        temperature:
          type: number
        top_p:
          type: number
    TextCompletionResponse:
      type: object
      required: [id, object, created, model, choices]
      properties:
        id:
          type: string
        object:
          type: string
          enum: [text_completion]
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              finish_reason:
                type: string
        usage:
          $ref: '#/components/schemas/Usage'
    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer